refactor: create authed route group
