+++ apps/nextjs/src/app/(authed)/feed/page.tsx
@@ -1,7 +1,7 @@
 import { Suspense } from "react";
 
 import { api } from "~/trpc/server";
-import { ReviewList } from "../../components/reviews";
+import { ReviewList } from "../../../components/reviews";
 
 export const runtime = "edge";
